Move ProgramOptions into a separate file
